MAINT: Check that all options are used
