+++ mne_bids_pipeline/tests/configs/config_ds000117.py
@@ -1,6 +1,5 @@
 """Faces dataset."""
 
-study_name = "ds000117"
 bids_root = "~/mne_data/ds000117"
 deriv_root = "~/mne_data/derivatives/mne-bids-pipeline/ds000117"
