+++ mne_bids_pipeline/tests/configs/config_ds000248_base.py
@@ -2,7 +2,6 @@
 
 import mne
 
-study_name = "ds000248"
 bids_root = "~/mne_data/ds000248"
 deriv_root = "~/mne_data/derivatives/mne-bids-pipeline/ds000248_base"
 subjects_dir = f"{bids_root}/derivatives/freesurfer/subjects"
