+++ mne_bids_pipeline/tests/configs/config_ds003104.py
@@ -1,6 +1,5 @@
 """Somato."""
 
-study_name = "MNE-somato-data-anonymized"
 bids_root = "~/mne_data/ds003104"
 deriv_root = "~/mne_data/derivatives/mne-bids-pipeline/ds003104"
 subjects_dir = f"{bids_root}/derivatives/freesurfer/subjects"
