+++ mne_bids_pipeline/tests/configs/config_ds003392.py
@@ -1,6 +1,5 @@
 """hMT+ Localizer."""
 
-study_name = "localizer"
 bids_root = "~/mne_data/ds003392"
 deriv_root = "~/mne_data/derivatives/mne-bids-pipeline/ds003392"
