+++ mne_bids_pipeline/tests/configs/config_ds003775.py
@@ -1,6 +1,5 @@
 """SRM Resting-state EEG."""
 
-study_name = "ds003775"
 bids_root = "~/mne_data/ds003775"
 deriv_root = "~/mne_data/derivatives/mne-bids-pipeline/ds003775"
