+++ mne_bids_pipeline/tests/configs/config_ds004107.py
@@ -9,9 +9,8 @@ International Congress Series, Volume 1300, Pages 615-618.
 
 # This has auditory, median, indx, visual, rest, and emptyroom but let's just
 # process the auditory (it's the smallest after rest)
-study_name = "ds004107"
-bids_root = f"~/mne_data/{study_name}"
-deriv_root = f"~/mne_data/derivatives/mne-bids-pipeline/{study_name}"
+bids_root = "~/mne_data/ds004107"
+deriv_root = "~/mne_data/derivatives/mne-bids-pipeline/ds004107"
 subjects = ["mind002"]
 sessions = ["01"]
 conditions = ["left", "right"]  # there are also tone and noise
