+++ mne_bids_pipeline/tests/configs/config_eeg_matchingpennies.py
@@ -1,6 +1,5 @@
 """Matchingpennies EEG experiment."""
 
-study_name = "eeg_matchingpennies"
 bids_root = "~/mne_data/eeg_matchingpennies"
 deriv_root = "~/mne_data/derivatives/mne-bids-pipeline/eeg_matchingpennies"
